+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/test.py
@@ -8,7 +8,7 @@ import click
 from .console import CONTEXT_SETTINGS, abort, echo_info, echo_success, echo_waiting
 from ..utils import running_on_ci
 from ..constants import get_root
-from ..test import construct_pytest_options, fix_coverage_report, get_tox_envs, pytest_coverage_sources
+from ..testing import construct_pytest_options, fix_coverage_report, get_tox_envs, pytest_coverage_sources
 from ...subprocess import run_command
 from ...utils import chdir, file_exists, remove_path
